Fix DAMS-NT Network DCPs for USACE NAE
A change for polled network DCPs broke the olde DAMS-NT netdcp interface that USACE NAE was using for Sutron DCPs. DAMS-NT network DCPs have the 38-byte GOES header. The new code was trying to parse the header as if it were a USGS EDL file. The fix differentiates between the two types of netdcp.