[build] added target to generate eclipse ide files.
It does pull in yet more dependencies when run but it was more direct
and obvious what was actually happening this way.

Creates the .project and .classpath file. It should be rerun if jars
are added or changed. Changes to .project should be included in
build.xml (or a file not in root created that this task can copy.)