Added ant tasks to build docs.
The target docs.build informs users of requirements if they aren't met
and failes graceful to html if PDFs can't be made.

Developer is informed of requirements to build PDFs in this case.

build.xml is in the docs directory for keeping it all clean.
Subant is not used at present since the primary build depends directly
on the tasks. This can change if we/when we split the opendcs.jar build
and installer creation into different folders and thus subprojects.

The ant contrib libraries include to allow for use of if in targets.