Corrections for different environments.
Set encoding on javac to UTF-8. There are comments/text with UTF8 characters that seemed reasonable to leave.

Altered build to download ant-junitlauncher if required.
The github action-runners appear to use a minimal install that doesn't include it. But keeping the version we
had in lib was causing version conflicts with no method found errors.