Balanced connection pool usage in RoutingScheduler.
RoutingSpecThread was pulling a connection but not closing it until the
loop finished.

The platformStatusDAI apparently was not able to properly close and
repoen it's connection during the loop operations.
The logic was changed to create,use, and close PlatformDAO each loop.
There is likely a minor performance hit to this, but that could be
solved in the future my keeping an instance for the makePlatformDAO
call.