Add MHD/T field types for changing msg timestamp.
This adds two new field types for the F operator: MHD and MHT
MHD is Message Header Date, and using that will change the message header date to the parsed value.
MHT is Message Header Time, and it works the same as MHD, but for the time portion instead.

These are needed as some message types, such as Iridium, contain the wrong time in the message header. Depending on congestion, the difference between the DCP transmission time and the time it is received can be anywhere from several seconds, to 10 minutes. For messages that also contain the interval and offset in the message, the offset is based on the message header time, however it's based on the transmission time, not receive time. This means trying to parse with F(MOFF) will cause the timestamps of the sensor values to be offset incorrectly. For a message received 5 minutes later, for example, it would store the value at :20, instead of :15.

These two new field types allow users to send the correct date and time in the message body itself, then override the header timestamp with these values, which will allow correct operation of the MOFF field type.

One item of note, is the Truncation operator: While the truncation operator actual modifies the message time in the same manner as MHD and MHT do now, MOFF ignores it, and always uses the raw message timestamp. This conflict had to be managed by storing two copies of the message time, the original (including if changed by MHD or MHT), as well as the post-truncation time. The "untruncated" time is only used by MOFF, as a replacement for it previously using the raw message time. This is what allows MHD and MHT to affect MOFF. The truncation operator interaction with MOFF has not changed from the perspective of the DECODES script. However using MHD or MHT after a truncation, will reset the truncation state, as if it never happened. Including the side affect of altering the message header time on first call. For this reason it's recommended to use MHD and MHT before any truncation operators.