Add support to change message header timestamp for Iridium parsing (#163)
* Add MHD/T field types for changing msg timestamp.

This adds two new field types for the F operator: MHD and MHT
MHD is Message Header Date, and using that will change the message header date to the parsed value.
MHT is Message Header Time, and it works the same as MHD, but for the time portion instead.

These are needed as some message types, such as Iridium, contain the wrong time in the message header. Depending on congestion, the difference between the DCP transmission time and the time it is received can be anywhere from several seconds, to 10 minutes. For messages that also contain the interval and offset in the message, the offset is based on the message header time, however it's based on the transmission time, not receive time. This means trying to parse with F(MOFF) will cause the timestamps of the sensor values to be offset incorrectly. For a message received 5 minutes later, for example, it would store the value at :20, instead of :15.

These two new field types allow users to send the correct date and time in the message body itself, then override the header timestamp with these values, which will allow correct operation of the MOFF field type.

One item of note, is the Truncation operator: While the truncation operator actual modifies the message time in the same manner as MHD and MHT do now, MOFF ignores it, and always uses the raw message timestamp. This conflict had to be managed by storing two copies of the message time, the original (including if changed by MHD or MHT), as well as the post-truncation time. The "untruncated" time is only used by MOFF, as a replacement for it previously using the raw message time. This is what allows MHD and MHT to affect MOFF. The truncation operator interaction with MOFF has not changed from the perspective of the DECODES script. However using MHD or MHT after a truncation, will reset the truncation state, as if it never happened. Including the side affect of altering the message header time on first call. For this reason it's recommended to use MHD and MHT before any truncation operators.

* Add tests for MHD and MHT.

The current testing framework had to be rewritten a bit to accommodate a different message type. Now it handles a few custom types: goes, iridium, and edl. The message type can be specified by append the type, in lower-case, to the input filename extension. The default is edl if not specified.

The following are valid input filenames:
TEST.input_iridium -> Iridium message
TEST.input_goes -> GOES message
TEST.input_data-logger -> EDL message
TEST.input_edl -> EDL message (alias for ease of use)
TEST.input -> EDL message (default)

More types can be added by modifying the appropriate functions in DecodesHelper.java.

* Update docs for new MHD and MHT field types.

Since there weren't any present so far, I decided on a naming scheme for documentation references. The naming should be globally unique (for OpenDCS at least), so the name is structured like this:

book.chapter.section.subsection

I used the headers as a breadcrumbs style navigation reference to construct these.
Such as: decodes.format.field.message-header

'decodes' for the 'DECODES Guide' book
'format' for 'DECODES Format Language' chapter
'field' for the 'Field Operations' section
'message-header' for the 'Message Header Fields' subsection under Field Operations.

I chose a one-two word name that best describes the appropriate part of each piece, so they're short, but still unique and descriptive enough.