Modified to allow heredoc to work.
Because of the use of a buffered reader for the tokens and a seperate
reader for the password it was impossible to get the password from a
pipe or heredoc.

Modified TTYEcho to take the buffered reader to maintain the input
stream. This allows a shell script to set the initial user in the docker
file.