Quick for XML decodes script processing.
The implementation leaves a bit to be desired, but it works. NOAA is
working on updates to their decoding so getting something in quick since
the XML processing is just broken.

Next steps are properly encapsulating the DecodesScript anyways (the
script was allowed to be unprepared by the time it was used to decode)
and this gave me a better indication of what to do.