Additional logging changes.
Added anonymous handler to pass log records to the QueueLogger for the AlarmHandler.

Longer term the AlarmHandler should be alter to take a more generic object so other implementations
could configure their loggers to pass message in for processing.

Alternatively the solution could be custom AlarmHandlers that are configured through the logging
implementation itself.