Initial generic value expander.
Creates a system such that users can configure values of propreties
with references to external systems. For example getting credentials
from the filesystem, or environment variables.

Implemented as a service provider so it is easier to expand in the future
for OpenDCS and advanced users of OpenDCS that have a specific need.