Connection pool and sql improvements.
Switch to a synchronized set for storing the connections as well as
add synchronized to a few members of the pool. It appears a lack of
synchronization was causing some bad behavior once more than one thread existed.

Cleaned up some sql usage, it appeared that some calls may not have been returning their connections
or were creating their own. It wasn't super clear, but switching to bind variables is a good thing anyways
and testing is showing correct behavior.

It appears that the SqlDatabaseIO::getConnection was getting called in a few places instead of CwmsSqlDatabaseIO
when the database was CWMS. Modified DataSourceExec to require a DataSource,Database constructor and adjusted
DataSource::makeDelegate to accomdate this change. This should help with making sure a given exec is always
using the same Database, and it helps march in the direction of Database not needing to exist as a singleton.