Reorganize integration test setup.
Additional tests will required judicious use of ExecutionConditions
to handle the multiple implementations and what is and isn't even available.

By moving the initial test setup, especially the Configuration class into an extension
we can store the instance in the Junit5 "Store" and the future ExecutionConditions
can find this instance and get relevant information about the system under test.