Add option to remove payload IE from msg data.
When the LRGS processes an Iridium message, it leaves the payload Information Element (IE) header in the content of the message itself. This requires the user to skip over it when using DECODES scripts.
It looks something like: "IE:020088 ".

Since this information is metadata, and not actually the payload content, it should be removed from the message data. However an option has been provided, which defaults to the existing behavior to avoid breaking existing DECODES scripts. The payload length was placed in the message header section, under PLEN, so it's not lost.