Add the new PLEN to IridiumPMParser. (#740)
This is an addition for missing code in PR #738. This adds the new PLEN
to the performance metrics parser, as well as fixes RAD, which broke in
the prior PR. This also fixes the header length in the decodes trace log
and edit window.

With this change, and changing the lrgs.conf setting, the first byte in
the message when running a DECODES script will be the first payload
byte. For Sutron SL3 short messages, this will be '0'.