Adapt analysis phase to work with external forks. (#849)
The current direct mechanism of analysis that we're using here and
rest_api is using doesn't work for PRs from a fork due to legitimate
security concerns over secrets.

After some research I found a way to enable the SonarCloud checks to
work for PRs from forks by splitting the analysis step into another
workflow that prevents PR code from getting executed with access to the
secrets.

This was tested in my own Fork to verify the initial behavior and may
require a few additional steps after merge to adjust any permissions.
(Sonar needs to be able to create and update comment with it's results.