Bugfix/779 makeconcrete tsdao usage (#797)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Fixes #779 . 

<!-- if you are referencing a specific issue a problem description is
not required -->
Describe the problem you are trying to solve.

## Solution

1. Change usages of DbCompResolver.makeConcrete to take a TimeSeriesDAI
instance so that it can be set to a transaction to avoid opening to many
connections in a tight loop.
2. Change from Oracle UCP in CWMS to HikariCP.

The change to makeConcrete usage *did* make the situtation better, but
we still ran out of connections. However, due to the extra layers of
abstraction from CwmsDbConnectionPool to Oracle UCP it is difficult to
determine if the opendcs CwmsConnectionPool wrapper is the main problem
or not. Since we're already planning to move to standard interfaces I've
decided to just try HikariCP here.

Initial results are good, and only affects CWMS.

I'm not sure this is the quite way to include a new connection pool
concept in the the software, but this problem is frustrating enough I'm
at least testing the "easy" way to verify were problems are.

## how you tested the change

Manually running compproc and seeing that it did fail in less than 10
minutes... and than 20 minutes.
About to start a long running test on our infrastructure.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.