Devops/add manual tests (#864)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Fixes #473 . 

## Solution

I created a plugin that processes Gherkin Feature files and translates
the data into Kiwi TCMS test cases.

Kiwi "Plans", which are groupings of test cases, are defined in the
build.gradle file of the integrationtesting/manual subproject.

Tests can be assigned to more than one plan. For example, we can have a
"test everything plan" and a "short smoke test" plan.

The use of Gherkin does add some complexity... I had to make an entire
gradle plugin and support library to actually use it. However, it
greatly simplifies structing tests which are repeated operations with
minor variations - best example, see the Platform Sorting by "one of the
4 options."

I created this because I was starting to write that out manually and
thus also started to go somewhat insane.

NOTE: focus is on Kiwi TCMS as HEC is self-hosting (no external access)
an instance. However the plugin and processing is designed to be
agnostic. It's definitely not there yet, but that's the goal. A later
version of the plugin could generate CSV files, PDF form, or even a
simple site (someone else is doing that work) if needed. Additionally
there are other TCMS systems so we'd likely work out a way to use them.

## how you tested the change

Currently manually tested from local machine to HEC's instance.
After merge and some initial testing, I will setup a link in an action
to push the tests continuously.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.

---------

Co-authored-by: Karl Tarbet <ktarbet@users.noreply.github.com>