Correct build artifact setup. (#975)
<!-- if you are referencing a specific issue a problem description is
not required -->
Validation on Maven Central requires the resulting pom.xml having a
project name and description.
As these are per artifact they must be set on the individual subprojects
instead of the template.

Additionally, due to the setup of the installer gradle attempts to make
a maven bundle to get pushed to maven central which we don't want or
need.

## Solution

* Correct publication type. (cherry picked from commit
70d5c3fa341c9c77709601e2e44c4cd75ed2aadf)
* Add publication information. (cherry picked from commit
62d840eb476fffb1bbd14b773f7f91ffaf351848)
* Prevent install bundle maven publications. (cherry picked from commit 

## how you tested the change

Changes were used on the 7.5 release branch.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.