Connection TimeZone bug fix (#1017)
change the default time zone of the conn to be UTC, and fix related bugs
in ResEvap.

## Problem Description

<!-- if you are referencing a specific issue a problem description is
not required -->
Describe the problem you are trying to solve.

The default connection timezone to the Oracle data base is set to the
timezone of the local machine. This causes querys which include
datetimes to print and return values assosiated with the incorrect time
zone.

## Solution

Describe your solution.

By unwrapping the connection within the getconnection function call we
can set it to have a new and correct defult time zone of UTC

## how you tested the change

Describe what was done to test the change. This section can be left
blank
if automated tests demonstrating usage are provided in the PR.

I was able to test the fix the viewing the results of the ResEvap
Algorithm, and varifying the calls to the database from that algorithm
returned the appropriate values.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.