Refactor/1012 cwms oracle migration updates (#1018)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Fixes #1012. 

<!-- if you are referencing a specific issue a problem description is
not required -->

## Solution

Transition to an "admin" (dba) account to create the schema instead of
making the CCP user ahead of time.

Also discovered an order of operations problem within the manageDatabase
application in regards to the configuration placeholder values.

## how you tested the change

The integration tests test this work.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.

---------

Co-authored-by: Karl Tarbet <ktarbet@users.noreply.github.com>