Feature/allow cwms provider to create user (#1058)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Fixes #. 

<!-- if you are referencing a specific issue a problem description is
not required -->
During initial setup work on our cloud provider it was discovered that
we weren't interacting with the database correctly.

## Solution

- Assume all operations are as the "builduser", which in the case of AWS
RDS is the default admin account that AWS will allow to be setup.
- CWMS Schema was changed to allow this user to bypass security checks;
previously it was SYS

## how you tested the change

The integration test setup covers this behavior.

NOTE: draft until the CWMS Database changes are merged in and deployed.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.

---------

Co-authored-by: Karl Tarbet <ktarbet@users.noreply.github.com>