Allow for difference in Oracle AWS RDS permissions assignment. (#1105)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Fixes #1103 . 

<!-- if you are referencing a specific issue a problem description is
not required -->
Describe the problem you are trying to solve.

## Solution

Detect if we are in an AWS RDS environment and call the appropriate
methods to assign the permissions need to the user.

## how you tested the change

Non-RDS is tested as part of the integration test flow.
RDS was tested manually to verify general logic was correct.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.