Task/1269 update logging part 16 (#1329)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Partial #1269. 

Depends on #1328 

<!-- if you are referencing a specific issue a problem description is
not required -->
Describe the problem you are trying to solve.

## Solution

Describe your solution.

## how you tested the change

Describe what was done to test the change. This section can be left
blank
if automated tests demonstrating usage are provided in the PR.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.