parameter column and property requirements (#1299)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->
Fixes #1286 and #1285

<!-- if you are referencing a specific issue a problem description is
not required -->
Describe the problem you are trying to solve.

makes it so you can set requiered proporties and see which role type of
the parameteres

## Solution

Add new Properties spec requiered and requirement group roles and added
an new column the the parameters table

## how you tested the change

Describe what was done to test the change. This section can be left
blank
if automated tests demonstrating usage are provided in the PR.

manual testing

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.