Updated lower and fix value check (#1451)
## Problem Description

<!-- if your PR does not address an open issue you can remove this line
-->

<!-- if you are referencing a specific issue a problem description is
not required -->
Describe the problem you are trying to solve.

Valid start and end time are generating aggregation periods outside of
the valid range, and algorithm testing are not checking the values
correctly in this range.

## Solution

Describe your solution.

The lower date range in the aggratiod period generation was not being
set correctly and the integration algorithm tests were looking for data
any outside fo the correct range.

## how you tested the change

Describe what was done to test the change. This section can be left
blank
if automated tests demonstrating usage are provided in the PR.

confirmed that the algorithm tests are now working as expected.

## Where the following done:

- [ ] Tests. Check all that apply:
   - [ ] Unit tests created or modified that run during ant test.
- [ ] Integration tests created or modified that run during integration
testing
         (Formerly called regression tests.)
   - [ ] Test procedure descriptions for manual testing
- [ ] Was relevant documentation updated?
- [ ] Were relevant config element (e.g. XML data) updated as
appropriate

If you aren't sure leave unchecked and we will help guide you to want
needs changing where.

---------

Co-authored-by: Oskar Hurst <oskar.a.hurst@usace.army.mil>