Task/1545 transaction context (#1550)
Fixes #1545.  Or at least starts in on it.

## Solution

1. Cleanup (remove) Injection of the OpenDcsDatabase instance into the
DAO (which made them stateful again)
2. Create a `TransactionContext` object that allows retrieving
appropriate runtime database/implementation information
3. Apply ServiceAnnotation and usage Lookup classes from
Netbeans/OpenIDE to retrieve the 3 DAOs adjusted. (CwmsLocationLevelDao
is an additional candidate but I didn't want to deal with that in what
is essentially a "demo/initial design PR".

Still need to account for differences in implementations which will
likely require some changes to the lookup method or some additional
annotations... or understanding of the "path" or other parameters to the
`ServiceAnnotation`. We will need to create a `UserManagementDao`
implementation for CWMS and we can address that behavior at that time.

## how you tested the change

No functional change. Existing Integration tests apply.

---------

Co-authored-by: Karl Tarbet <ktarbet@users.noreply.github.com>