Devops/swap library (#1582)
## Problem Description

An instance of, at least compproc at least for CWMS, running in a
container with `--cpus="1"` will pegg the cpu usage at 100%.

## Solution

Issue was tracked down to this line:
https://github.com/HydrologicEngineeringCenter/hec-service-annotations/blob/337c0f4d1082ca6c121204665a21a4cd111105c9/lookup/src/main/java/mil/army/usace/hec/util/lookup/implspi/ActiveQueue.java#L56,
specifically the run function calls remove and the linked line is what
is executed, handler in run just continues the while loop if that
exception is thrown.

I've now excluded the old hec/rma Lookup libraries and implemented the
new compatibility layer (we're already using the newer Netbeans lookup
for the DAOs).

The CPU usage became < 1% in the same conditions.

## how you tested the change

- Existing integration tests to verify previous lookup behavior (the
CWMS Implementation usage classes annotated with the RMA namespaced
ServiceAnnotatiosn)
- manually in a docker container with cpus=1 and memory=2048 to mimic a
simple and common AWS Fargate setup. Inspected with VisualVM to see that
the Thread that was in a spin loop now sits WAITING.